+++ src/graph_notebook/magics/metadata.py
@@ -204,8 +204,11 @@ def build_gremlin_metadata_from_query(query_type: str, results: any, res: Respon
     if query_type == 'explain':
         gremlin_metadata = create_propertygraph_metadata_obj('explain')
         gremlin_metadata.set_request_metrics(res)
-        gremlin_metadata.set_metric_value('predicates', int((re.search(r'# of predicates: (.*?)\n', results).group(1))
-                                                            .replace(".", '').replace(",", '')))
+        try:
+            gremlin_metadata.set_metric_value('predicates', int((re.search(r'# of predicates: (.*?)\n', results).group(1))
+                                                                .replace(".", '').replace(",", '')))
+        except AttributeError:
+            pass
         return gremlin_metadata
     elif query_type == 'profile':
         gremlin_metadata = create_propertygraph_metadata_obj('profile')